+++ src/main/java/com/goormitrip/goormitrip/user/service/UserDetailsServiceImpl.java
@@ -5,6 +5,8 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
+import com.goormitrip.goormitrip.global.security.CustomUserDetails;
+import com.goormitrip.goormitrip.user.domain.UserEntity;
 import com.goormitrip.goormitrip.user.repository.UserRepository;
 
 @Service
@@ -18,8 +20,8 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
 	@Override
 	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-		return userRepository.findByEmail(email)
+		final UserEntity user = userRepository.findByEmail(email)
 			.orElseThrow(() -> new UsernameNotFoundException("이메일 또는 비밀번호를 확인해주세요."));
+		return new CustomUserDetails(user);
 	}
-
 }
